+++ src/rust/crates/ain-evm-ffi/src/lib.rs
@@ -34,31 +34,31 @@ mod ffi {
 }
 
 pub fn evm_add_balance(address: &str, amount: [u8; 32]) -> Result<(), Box<dyn Error>> {
-    RUNTIME.evm.add_balance(address, amount.into())
+    RUNTIME.handlers.evm.add_balance(address, amount.into())
 }
 
 pub fn evm_sub_balance(address: &str, amount: [u8; 32]) -> Result<(), Box<dyn Error>> {
-    RUNTIME.evm.sub_balance(address, amount.into())
+    RUNTIME.handlers.evm.sub_balance(address, amount.into())
 }
 
 pub fn evm_validate_raw_tx(tx: &str) -> Result<bool, Box<dyn Error>> {
-    match RUNTIME.evm.validate_raw_tx(tx) {
+    match RUNTIME.handlers.evm.validate_raw_tx(tx) {
         Ok(_) => Ok(true),
         Err(_) => Ok(false),
     }
 }
 
 pub fn evm_get_context() -> u64 {
-    RUNTIME.evm.get_context()
+    RUNTIME.handlers.evm.get_context()
 }
 
 fn evm_discard_context(context: u64) {
     // TODO discard
-    RUNTIME.evm.discard_context(context)
+    RUNTIME.handlers.evm.discard_context(context)
 }
 
 fn evm_queue_tx(context: u64, raw_tx: &str) -> Result<bool, Box<dyn Error>> {
-    match RUNTIME.evm.queue_tx(context, raw_tx) {
+    match RUNTIME.handlers.evm.queue_tx(context, raw_tx) {
         Ok(_) => Ok(true),
         Err(_) => Ok(false),
     }
@@ -66,6 +66,6 @@ fn evm_queue_tx(context: u64, raw_tx: &str) -> Result<bool, Box<dyn Error>> {
 
 use rlp::Encodable;
 fn evm_finalise(context: u64, update_state: bool) -> Result<Vec<u8>, Box<dyn Error>> {
-    let (block, _failed_tx) = RUNTIME.evm.finalize_block(context, update_state)?;
+    let (block, _failed_tx) = RUNTIME.handlers.evm.finalize_block(context, update_state)?;
     Ok(block.header.rlp_bytes().into())
 }
