+++ src/rust/crates/ain-grpc/Cargo.toml
@@ -23,6 +23,7 @@ serde_json = "1.0"
 tokio = { version = "1.1", features = ["rt-multi-thread"] }
 tonic = "0.8"
 primitive-types = "0.12.1"
+ethereum = "0.14.0"
 
 [build-dependencies]
 cxx-gen = "0.7"
