+++ src/rust/protobuf/rpc/eth.proto
@@ -5,16 +5,18 @@ import "google/protobuf/empty.proto";
 import "types/eth.proto";
 
 service eth {
-    /// Returns eth_accounts list.
-    rpc Eth_Accounts(google.protobuf.Empty) returns (types.EthAccountsResult);
+  /// Returns eth_accounts list.
+  rpc Eth_Accounts(google.protobuf.Empty) returns (types.EthAccountsResult);
 
-    /// Call contract, returning the output data. Does not create a transaction.
-	rpc Eth_Call(types.EthCallInput) returns (types.EthCallResult);
+  /// Call contract, returning the output data. Does not create a transaction.
+  rpc Eth_Call(types.EthCallInput) returns (types.EthCallResult);
 
-    /// Returns the balance for the given address.
-	rpc Eth_GetBalance(types.EthGetBalanceInput) returns (types.EthGetBalanceResult);
+  /// Returns the balance for the given address.
+  rpc Eth_GetBalance(types.EthGetBalanceInput) returns (types.EthGetBalanceResult);
 
-    /// [ignore]
-    /// Returns the balance for the given address.
-	rpc Eth_SendTransaction(types.EthSendTransactionInput) returns (types.EthSendTransactionResult);
+  rpc Eth_GetBlockByHash(types.EthGetBlockByHashInput) returns (types.EthGetBlockByHashResult);
+
+  /// [ignore]
+  /// Returns the balance for the given address.
+  rpc Eth_SendTransaction(types.EthSendTransactionInput) returns (types.EthSendTransactionResult);
 }
